refactor: template behavior gets its own file
